+++ addons/mgsr_vests/UI/gear_mgsr_vest_02_x_ca.paa
