+++ addons/mgsr_vests/UI/gear_mgsr_vest_03_x_ca.paa
