+++ addons/mgsr_vests/data/mgsr_belt.rvmat
