+++ addons/mgsr_vests/data/mgsr_belt_smdi.paa
