+++ addons/mgsr_vests/data/mgsr_vest.rvmat
