+++ addons/mgsr_vests/data/mgsr_vest_01_co.paa
