+++ addons/mgsr_vests/data/mgsr_vest_02_co.paa
