+++ addons/mgsr_vests/data/mgsr_vest_03_co.paa
