+++ addons/mgsr_vests/data/mgsr_vest_smdi.paa
