+++ addons/mgsr_vests/mgsr_vest.p3d
