+++ addons/mgsr_vests/model.cfg
@@ -151,5 +151,6 @@ class CfgModels
         };
         skeletonName = "OFP2_ManSkeleton";
     };
+    class mgsr_vest: ArmaMan {};
     class mgsr_carrier_belt: ArmaMan {};
 };
